Improve session model relations to car and track models
This commit changes the way tracks and cars are related to sessions. The
implementation now relies on car and track names, along with a season id,
to accurately relate a standalone car or track model.

Previously, the relationships between models were declared as follows:
- Race has_one Track
- RacerEntry has_one Car

The changes introduced in this commit modify these Id-based relations into
simple parameters of type String corresponding to Track and Car's names,
respectively.

The match between these names and their corresponding models no longer takes
place when constructing the Session model upon import, but rather when RVA
results are calculated over at rva_calculate_results_service.rb. Here, they
are queried alongside the related session's season for a model match. For
example:

  Track.find { |t| t.name.eql?(track_name) && t.season.eql?(@session.season) }

In case a model match isn't found, then it can be handled and displayed in
the results table as 'not found' or as just the name itself, instead of
breaking the entire sessions' view.

Because the searches are no longer based on IDs, this change also allows for
car and track models to be freely deleted and re-added to the site's database.